refactoring and creating a repository for the connection with databse
