+++ src/controllers/users.controller.js
@@ -1,5 +1,8 @@
-import connection from "../database/database.js";
-import { listUsers } from "../repositories/users.repositories.js";
+import {
+	listUserPosts,
+	listUsers,
+	listUsersById,
+} from "../repositories/users.repositories.js";
 import {
 	badRequestResponse,
 	serverErrorResponse,
@@ -12,28 +15,13 @@ async function getUserInfo(req, res) {
 		if (!id) {
 			badRequestResponse(res);
 		}
-		const userExists = await connection.query(
-			`
-      SELECT * FROM users WHERE id = $1;
-    `,
-			[id]
-		);
+		const userExists = await listUsersById(id);
 		if (userExists.rowCount === 0) {
 			return res.sendStatus(404);
 		}
 
-		const userInfo = await connection.query(
-			`
-        SELECT 
-            users.id, users.username, users.pictureurl,
-            json_agg(posts.*) AS posts
-        FROM users
-        JOIN posts ON users.id = posts.userid
-        WHERE users.id = $1
-        GROUP BY (users.id)
-        ;`,
-			[id]
-		);
+		const userInfo = await listUserPosts(id);
+
 		return res.status(STATUS_CODE.OK).send(userInfo.rows[0]);
 	} catch (error) {
 		serverErrorResponse(res, error);
